requirements.txt
streamlit==0.89.0
keras==2.6.0
numpy==1.19.5
tensorflow==2.6.0
pandas==1.3.3
Pillow==8.3.2
